refactor(cli): migrate ghsl tooling onto discovery
Move ghsl's prober/validator/live_validator/tile_regen to earthlens.ghsl.cli,
wiring the first tile_regen role through discovery (refresh's _TILE_REGENS projects
dispatch_table). Add http_head to the public toolkit. Relocate ghsl's tile-regen
and live tests to earthlens-land; repoint the core --tiles command test. Shrink
the guard allow-lists (ghsl; ghsl import).
